# trailmemo/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_trailmemo_pkg_response.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error_code:
        example: ROUTE_NOT_FOUND
        type: string
      message:
        example: success
        type: string
      request_id:
        type: string
    type: object
  internal_handler.ChangePasswordRequest:
    properties:
      newPassword:
        example: "654321"
        maxLength: 128
        minLength: 6
        type: string
      oldPassword:
        example: "123456"
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  internal_handler.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: testuser
        type: string
    required:
    - password
    - username
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: "123456"
        maxLength: 128
        minLength: 6
        type: string
      phone:
        example: "+8613800138000"
        type: string
      username:
        example: testuser
        maxLength: 64
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  internal_handler.UpdateUserInfoRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      nickname:
        example: 新昵称
        type: string
    type: object
  internal_handler.WechatLoginRequest:
    properties:
      code:
        example: 081aBc1234567
        type: string
    required:
    - code
    type: object
host: localhost:8087
info:
  contact:
    email: support@example.com
    name: API Support
  description: 迹忆旅图寻迹旅游后端 API 文档
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://example.com/terms
  title: TrailMemo API
  version: 1.0.0
paths:
...
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 用户登录
      tags:
      - 用户
  /api/v1/users/login/wechat:
    post:
      consumes:
      - application/json
      description: 使用微信授权码登录
      parameters:
      - description: 微信授权码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.WechatLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 微信小程序登录
      tags:
      - 用户
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的昵称和头像
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateUserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
  /api/v1/users/me/password:
    put:
      consumes:
      - application/json
      description: 修改当前登录用户的密码
      parameters:
      - description: 密码信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - 用户
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 用户注册
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"